move tests that dont use network / unit tests to own file
